+++ nilearn/plotting/html_stat_map.py
@@ -15,12 +15,12 @@ from nibabel.affines import apply_affine
 from nilearn import DEFAULT_DIVERGING_CMAP
 from nilearn._utils import check_niimg_3d, fill_doc
 from nilearn._utils.extmath import fast_abs_percentile
+from nilearn._utils.html_document import HTMLDocument
 from nilearn._utils.niimg import safe_get_data
 from nilearn._utils.param_validation import check_threshold
 from nilearn.datasets import load_mni152_template
 from nilearn.image import get_data, new_img_like, reorder_img, resample_to_img
 from nilearn.plotting.find_cuts import find_xyz_cut_coords
-from nilearn.plotting.html_document import HTMLDocument
 from nilearn.plotting.img_plotting import load_anat
 from nilearn.plotting.js_plotting_utils import colorscale, get_html_template
