+++ nilearn/plotting/js_plotting_utils.py
@@ -12,11 +12,11 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from nilearn._utils.extmath import fast_abs_percentile
-from nilearn._utils.param_validation import check_threshold
-from nilearn.plotting.html_document import (  # noqa: F401
+from nilearn._utils.html_document import (  # noqa: F401
     HTMLDocument,
     set_max_img_views_before_warning,
 )
+from nilearn._utils.param_validation import check_threshold
 from nilearn.surface import load_surf_mesh
 
 MAX_IMG_VIEWS_BEFORE_WARNING = 10
