+++ nilearn/plotting/tests/test_html_document.py
@@ -6,7 +6,7 @@ import pytest
 import requests
 from numpy.testing import assert_no_warnings
 
-from nilearn.plotting import html_document
+from nilearn._utils import html_document
 
 # Note: html output by nilearn view_* functions
 # should validate as html5 using https://validator.w3.org/nu/ with no
